client: first pass updating web3.py
