+++ client/setup.py
@@ -25,37 +25,17 @@ setup(
         "mypy==0.790",
         "mypy-protobuf==1.23",
         "flake8==3.8.3",
-        "pylint==2.6",
+        "pylint",
+        # "pylint>=2",
         "click==7.0",
         "click-default-group==1.2",
-        "attrdict==2.0.1",
-        "certifi==2018.11.29",
-        "chardet==3.0.4",
-        "cytoolz==0.11.0",
-        "eth-abi==1.3.0",
-        "eth-account==0.3.0",
-        "eth-hash==0.2.0",
-        "eth-keyfile==0.5.1",
-        "eth-keys==0.2.1",
-        "eth-rlp==0.1.2",
-        "eth-typing==2.1.0",
-        "eth-utils==1.4.1",
         "grpcio==1.33.2",
         "grpcio-tools==1.33.2",
-        "hexbytes==0.1.0",
-        "idna==2.8",
-        "lru-dict==1.1.6",
-        "parsimonious==0.8.1",
         "protobuf==3.13.0",
         "py_ecc==1.7.1",
         "py-solc-x==1.1.0",
-        "pycryptodome==3.9.8",
         "cryptography==3.3.2",
-        "rlp==1.1.0",
-        "semantic-version==2.8.4",
-        "six==1.15.0",
-        "toolz==0.9.0",
-        "web3>=4,<5",
+        "web3>=5<6",
     ],
     entry_points={
         'console_scripts': [
