+++ client/zeth/core/contracts.py
@@ -7,11 +7,10 @@
 from __future__ import annotations
 from zeth.core.utils import EtherValue
 from zeth.core.constants import SOL_COMPILER_VERSION
-from web3.utils.contracts import find_matching_event_abi  # type: ignore
-from web3.utils.events import get_event_data  # type: ignore
-from eth_utils import event_abi_to_log_topic  # type: ignore
+from web3._utils.contracts import find_matching_event_abi
+from eth_utils import event_abi_to_log_topic
 import solcx
-from typing import Dict, List, Iterator, Optional, Union, Iterable, Any
+from typing import Dict, List, Iterator, Optional, Union, Iterable, Any, cast
 
 # Avoid trying to read too much data into memory
 SYNC_BLOCKS_PER_BATCH = 1000
@@ -202,10 +201,12 @@ def get_event_logs(
     # skpping events with other topics, from the same contract.
 
     contract_address = instance.address
-    event_abi = find_matching_event_abi(instance.abi, event_name=event_name)
-    log_topic = event_abi_to_log_topic(event_abi)
-    batch_size = batch_size or SYNC_BLOCKS_PER_BATCH
+    contract_event = instance.events[event_name]()
 
+    event_abi = find_matching_event_abi(instance.abi, event_name=event_name)
+    log_topic = event_abi_to_log_topic(cast(Dict[str, Any], event_abi))
+
+    batch_size = batch_size or SYNC_BLOCKS_PER_BATCH
     while start_block <= end_block:
         # Filters are *inclusive* wrt "toBlock", hence the -1 here, and +1 to
         # set start_block before iterating.
@@ -218,12 +219,12 @@ def get_event_logs(
         logs = web3.eth.getLogs(filter_params)
         for log in logs:
             if log_topic == log['topics'][0]:
-                yield get_event_data(event_abi, log)
+                yield contract_event.processLog(log)
         start_block = to_block + 1
 
 
 def get_event_logs_from_tx_receipt(
-        instance_desc: InstanceDescription,
+        instance: Any,
         event_name: str,
         tx_receipt: Any) -> Iterator[Any]:
     """
@@ -232,9 +233,11 @@ def get_event_logs_from_tx_receipt(
     objects to be decoded by the caller. This function intentionally avoids
     connecting to a node, or creating host-side filters.
     """
-    contract_address = instance_desc.address
-    event_abi = find_matching_event_abi(instance_desc.abi, event_name=event_name)
-    log_topic = event_abi_to_log_topic(event_abi)
+    contract_address = instance.address
+    contract_event = instance.events[event_name]()
+
+    event_abi = find_matching_event_abi(instance.abi, event_name=event_name)
+    log_topic = event_abi_to_log_topic(cast(Dict[str, Any], event_abi))
     for log in tx_receipt.logs:
         if log.address == contract_address and log_topic == log['topics'][0]:
-            yield get_event_data(event_abi, log)
+            yield contract_event.processLog(log)
